+++ libs/core/src/index.ts
@@ -1,2 +1,3 @@
 export * as Community from './community';
+export * from './shared-types';
 export * from './types';
