+++ packages/commonwealth/client/scripts/controllers/server/roles.ts
@@ -1,6 +1,5 @@
+import { AccessLevel } from '@hicommonwealth/core';
 import app from 'state';
-
-import { AccessLevel } from 'permissions';
 import Account from '../../models/Account';
 import AddressInfo from '../../models/AddressInfo';
 import RoleInfo from '../../models/RoleInfo';
