+++ packages/commonwealth/client/scripts/views/components/Comments/CreateComment.tsx
@@ -1,15 +1,13 @@
-import React, { useEffect, useMemo, useState } from 'react';
-
+import { ContentType } from '@hicommonwealth/core';
 import { notifyError } from 'controllers/app/notifications';
-import type { DeltaStatic } from 'quill';
-import Thread from '../../../models/Thread';
-
 import { SessionKeyError } from 'controllers/server/sessions';
 import { useDraft } from 'hooks/useDraft';
+import type { DeltaStatic } from 'quill';
+import React, { useEffect, useMemo, useState } from 'react';
 import app from 'state';
 import { useCreateCommentMutation } from 'state/api/comments';
-import { ContentType } from 'types';
 import { useSessionRevalidationModal } from 'views/modals/SessionRevalidationModal';
+import Thread from '../../../models/Thread';
 import { jumpHighlightComment } from '../../pages/discussions/CommentTree/helpers';
 import { createDeltaFromText, getTextFromDelta } from '../react_quill_editor';
 import { serializeDelta } from '../react_quill_editor/utils';
