+++ packages/commonwealth/client/scripts/views/pages/notifications/helpers.tsx
@@ -1,13 +1,10 @@
+import type { IForumNotificationData } from '@hicommonwealth/core';
+import { NotificationCategories, ProposalType } from '@hicommonwealth/core';
+import { pluralize } from 'helpers';
 import _ from 'lodash';
 import moment from 'moment';
-import React from 'react';
-
 import 'pages/notifications/notification_row.scss';
-
-import { NotificationCategories, ProposalType } from '@hicommonwealth/core';
-import type { IForumNotificationData } from 'types';
-
-import { pluralize } from 'helpers';
+import React from 'react';
 import app from 'state';
 import { getCommunityUrl, getThreadUrl } from 'utils';
 import { User } from 'views/components/user/user';
