+++ packages/commonwealth/server/controllers/server_notifications_methods/emit.ts
@@ -1,7 +1,7 @@
+import { NotificationDataAndCategory } from '@hicommonwealth/core';
+import { NotificationInstance } from '../../models/notification';
 import emitNotifications from '../../util/emitNotifications';
 import { ServerNotificationsController } from '../server_notifications_controller';
-import { NotificationInstance } from '../../models/notification';
-import { NotificationDataAndCategory } from 'types';
 
 export type EmitOptions = {
   notification: NotificationDataAndCategory;
@@ -13,12 +13,12 @@ export type EmitResult = NotificationInstance;
 
 export async function __emit(
   this: ServerNotificationsController,
-  { notification, excludeAddresses, includeAddresses }: EmitOptions
+  { notification, excludeAddresses, includeAddresses }: EmitOptions,
 ): Promise<EmitResult> {
   return emitNotifications(
     this.models,
     notification,
     excludeAddresses,
-    includeAddresses
+    includeAddresses,
   );
 }
