+++ packages/commonwealth/server/routes/startEmailLogin.ts
@@ -3,11 +3,10 @@ import {
   formatFilename,
   loggerFactory,
 } from '@hicommonwealth/adapters';
-import { WalletId } from '@hicommonwealth/core';
+import { DynamicTemplate, WalletId } from '@hicommonwealth/core';
 import sgMail from '@sendgrid/mail';
 import type { NextFunction, Request, Response } from 'express';
 import moment from 'moment';
-import { DynamicTemplate } from '../../shared/types';
 import {
   LOGIN_RATE_LIMIT_MINS,
   LOGIN_RATE_LIMIT_TRIES,
