+++ packages/commonwealth/server/scripts/emails.ts
@@ -1,23 +1,20 @@
-import { Label as ChainEventLabel } from '../../shared/chain/labelers/util';
-import type { CWEvent } from '../../shared/chain/types/types';
-
 import { formatFilename, loggerFactory } from '@hicommonwealth/adapters';
-import { NotificationCategories } from '@hicommonwealth/core';
-import { capitalize } from 'lodash';
-import { Op, WhereOptions } from 'sequelize';
 import type {
   IChainEventNotificationData,
   IForumNotificationData,
   ISnapshotNotificationData,
-} from '../../shared/types';
+} from '@hicommonwealth/core';
+import { DynamicTemplate, NotificationCategories } from '@hicommonwealth/core';
+import { capitalize } from 'lodash';
+import { Op, WhereOptions } from 'sequelize';
+import { Label as ChainEventLabel } from '../../shared/chain/labelers/util';
+import type { CWEvent } from '../../shared/chain/types/types';
 import {
   formatAddressShort,
   getThreadUrl,
   renderQuillDeltaToText,
   smartTrim,
 } from '../../shared/utils';
-
-import { DynamicTemplate } from '../../shared/types';
 import { SENDGRID_API_KEY } from '../config';
 import { DB } from '../models';
 import { AddressAttributes } from '../models/address';
