+++ packages/commonwealth/server/util/webhooks/getWebhookData.ts
@@ -1,7 +1,9 @@
-import { NotificationCategories } from '@hicommonwealth/core';
+import {
+  NotificationCategories,
+  NotificationDataAndCategory,
+} from '@hicommonwealth/core';
 import { capitalize } from 'lodash';
 import { Label as chainEventLabel } from '../../../shared/chain/labelers/util';
-import { NotificationDataAndCategory } from '../../../shared/types';
 import { renderQuillDeltaToText, smartTrim } from '../../../shared/utils';
 import { SERVER_URL } from '../../config';
 import { CommunityInstance } from '../../models/community';
