+++ packages/commonwealth/server/util/webhooks/util.ts
@@ -1,7 +1,9 @@
 import { formatFilename, loggerFactory } from '@hicommonwealth/adapters';
-import { NotificationCategories } from '@hicommonwealth/core';
+import {
+  NotificationCategories,
+  NotificationDataAndCategory,
+} from '@hicommonwealth/core';
 import { Op } from 'sequelize';
-import { NotificationDataAndCategory } from '../../../shared/types';
 import { slugify } from '../../../shared/utils';
 import { DEFAULT_COMMONWEALTH_LOGO, SERVER_URL } from '../../config';
 import models from '../../database';
