+++ packages/commonwealth/server/workers/commonwealthConsumer/messageProcessors/snapshotConsumer.ts
@@ -3,9 +3,11 @@ import {
   RmqSnapshotNotification,
   StatsDController,
 } from '@hicommonwealth/adapters';
-import { NotificationCategories } from '@hicommonwealth/core';
+import {
+  NotificationCategories,
+  SnapshotEventType,
+} from '@hicommonwealth/core';
 import axios from 'axios';
-import { SnapshotEventType } from 'types';
 import type { DB } from '../../../models';
 import emitNotifications from '../../../util/emitNotifications';
