--- packages/commonwealth/shared/permissions.ts
@@ -1,6 +1,0 @@
-export enum AccessLevel {
-  Admin = 'admin',
-  Moderator = 'moderator',
-  Member = 'member',
-  Everyone = 'everyone',
-}
