+++ packages/commonwealth/test/unit/subscriptions/subscriptionMapping.spec.ts
@@ -1,11 +1,12 @@
-import { NotificationCategories, SupportedNetwork } from '@hicommonwealth/core';
-import chai from 'chai';
-import chaiHttp from 'chai-http';
 import {
   IBaseForumNotificationData,
+  NotificationCategories,
   NotificationDataAndCategory,
   SnapshotEventType,
-} from 'types';
+  SupportedNetwork,
+} from '@hicommonwealth/core';
+import chai from 'chai';
+import chaiHttp from 'chai-http';
 import { mapNotificationsDataToSubscriptions } from '../../../server/util/subscriptionMapping';
 
 chai.use(chaiHttp);
